+++ x/perp/keeper/grpc_query_test.go
@@ -98,6 +98,7 @@ func TestQueryPosition(t *testing.T) {
 				/* baseReserve */ tc.baseAssetReserve,
 				/* fluctuationLimitRatio */ sdk.OneDec(),
 				/* maxOracleSpreadRatio */ sdk.OneDec(),
+				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 			)
 			perpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
 				Pair: common.PairBTCStable,