+++ x/perp/keeper/liquidate.go
@@ -57,7 +57,9 @@ func (k Keeper) Liquidate(
 	}
 
 	params := k.GetParams(ctx)
-	err = requireMoreMarginRatio(marginRatio, params.MaintenanceMarginRatio, false)
+
+	maintenanceMarginRatio := k.VpoolKeeper.GetMaintenanceMarginRatio(ctx, pair)
+	err = requireMoreMarginRatio(marginRatio, maintenanceMarginRatio, false)
 	if err != nil {
 		return sdk.Coin{}, sdk.Coin{}, types.ErrMarginHighEnough
 	}