+++ x/perp/types/expected_keepers.go
@@ -260,6 +260,7 @@ type VpoolKeeper interface {
 	) (price sdk.Dec, err error)
 
 	IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) bool
+	GetMaintenanceMarginRatio(ctx sdk.Context, pair common.AssetPair) sdk.Dec
 	// ExistsPool returns true if pool exists, false if not.
 	ExistsPool(ctx sdk.Context, pair common.AssetPair) bool
 	GetSettlementPrice(ctx sdk.Context, pair common.AssetPair) (sdk.Dec, error)