+++ x/testutil/mock/perp_interfaces.go
@@ -450,6 +450,20 @@ func (mr *MockVpoolKeeperMockRecorder) GetCurrentTWAP(arg0, arg1 interface{}) *g
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetCurrentTWAP", reflect.TypeOf((*MockVpoolKeeper)(nil).GetCurrentTWAP), arg0, arg1)
 }
 
+// GetMaintenanceMarginRatio mocks base method.
+func (m *MockVpoolKeeper) GetMaintenanceMarginRatio(arg0 types2.Context, arg1 common.AssetPair) types2.Dec {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetMaintenanceMarginRatio", arg0, arg1)
+	ret0, _ := ret[0].(types2.Dec)
+	return ret0
+}
+
+// GetMaintenanceMarginRatio indicates an expected call of GetMaintenanceMarginRatio.
+func (mr *MockVpoolKeeperMockRecorder) GetMaintenanceMarginRatio(arg0, arg1 interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMaintenanceMarginRatio", reflect.TypeOf((*MockVpoolKeeper)(nil).GetMaintenanceMarginRatio), arg0, arg1)
+}
+
 // GetQuoteAssetPrice mocks base method.
 func (m *MockVpoolKeeper) GetQuoteAssetPrice(arg0 types2.Context, arg1 common.AssetPair, arg2 types1.Direction, arg3 types2.Dec) (types2.Dec, error) {
 	m.ctrl.T.Helper()