+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#730](https://github.com/NibiruChain/nibiru/pull/730) - update localnet script.
 * [#736](https://github.com/NibiruChain/nibiru/pull/736) - Bumps [github.com/spf13/cast](https://github.com/spf13/cast) from 1.4.1 to 1.5.0
 * [#735](https://github.com/NibiruChain/nibiru/pull/735) - Bump github.com/spf13/cobra from 1.4.0 to 1.5.0
+* [#729](https://github.com/NibiruChain/nibiru/pull/729) - move maintenance margin to the vpool module
 
 ### API Breaking